backend/requirements.txt
PyMuPDF==1.16.14
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
llama-cloud-services==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
google-generativeai==0.3.0